# db: add Iterator.RangeKeyChanged